move the combine action to a later stage in the build process
This change moves the actual combining of the content of files to the
moment the content is actually requested. This drastically reduces the
start up time of the BT, with a smaller penalty of a few seconds on the
first load of the app in dev mode. Can most likely be further improved.